driver-geominimal-auto-suggestions/requirements.txt
git+https://gitlab+deploy-token-75:hAoek7ygRC_nNkKN83dX@gitlab.city-srv.ru/public_city/citymobil_python_clickhouse_wrapper.git@0.2.4#egg=citymobil_python_clickhouse_wrapper
git+https://gitlab+deploy-token-83:UVXbZBQUGhLdTgxsxtxy@gitlab.city-srv.ru/public_city/citymobil-python-mysql-wrapper.git@0.2.0#egg=citymobil_python_mysql_wrapper
pandas~=1.3.4
pandahouse~=0.2.7
h3~=3.7.3
numpy~=1.21.4
geopandas~=0.10.2
sklearn~=0.0
scikit-learn~=1.0.1
Shapely~=1.8.0
geodaisy~=0.1.1
python-dotenv~=0.19.2
